refactor into llm module, use PyResult
